menu.ts is refactored - services classes are extracted
