+++ src/git.ts
@@ -13,8 +13,8 @@ export class Git {
     generateCommitMessage = async (): Promise<void> => {
         let chatUrl = this.app.configuration.endpoint_chat
         if (!chatUrl) chatUrl = this.app.configuration.endpoint_tools;
-        let chatModel = this.app.menu.getChatModel();    
-        if (!this.app.menu.isChatModelSelected()) chatModel = this.app.menu.getToolsModel();
+        let chatModel = this.app.getChatModel();    
+        if (!this.app.isChatModelSelected()) chatModel = this.app.getToolsModel();
         if (chatModel.endpoint) {
             const chatEndpoint = Utils.trimTrailingSlash(chatModel.endpoint)
             chatUrl = chatEndpoint ? chatEndpoint + "/" : "";
@@ -22,7 +22,7 @@ export class Git {
         if (!chatUrl) {
             const shouldSelectModel = await Utils.showUserChoiceDialog("Select a chat or tools model or an env with chat or tools model to generate a commit message.","Select")
             if (shouldSelectModel){
-                this.app.menu.showEnvView();
+                this.app.llamaWebviewProvider.showEnvView();
                 vscode.window.showInformationMessage("After the chat/tools model is loaded, try again generating commit message.")
                 return;
             }
