+++ src/services/external-model-strategy.ts
@@ -1,14 +1,7 @@
 import * as vscode from "vscode";
-import { QuickPickItem } from "vscode";
 import { Application } from "../application";
 import { IAddStrategy, LlmModel, ModelTypeDetails } from "../types";
 import { Utils } from "../utils";
-import * as axios from "axios";
-import { ModelType, UI_TEXT_KEYS, HF_MODEL_TEMPLATES, SETTING_TO_MODEL_TYPE, MODEL_TYPE_CONFIG } from "../constants";
-import * as path from "path";
-import * as fs from "fs";
-import { Configuration } from "../configuration";
-
 
 export class ExternalModelStrategy implements IAddStrategy {
     private app: Application;
