+++ src/services/local-model-strategy.ts
@@ -1,12 +1,8 @@
 
 import * as vscode from "vscode";
-import { QuickPickItem } from "vscode";
 import { Application } from "../application";
 import { IAddStrategy, LlmModel, ModelTypeDetails } from "../types";
 import { Utils } from "../utils";
-import * as axios from "axios";
-import { ModelType, UI_TEXT_KEYS, HF_MODEL_TEMPLATES, SETTING_TO_MODEL_TYPE, MODEL_TYPE_CONFIG } from "../constants";
-
 
 export class LocalModelStrategy implements IAddStrategy {
     private app: Application;
