+++ src/utils.ts
@@ -1,4 +1,4 @@
-import vscode, { Uri } from "vscode";
+import vscode, { QuickPickItem, Uri } from "vscode";
 import { exec } from 'child_process';
 import fs from 'fs';
 import path from 'path';
@@ -847,4 +847,17 @@ export class Utils {
         vscode.window.showErrorMessage(`Maximum attempts (${maxAttempts}) reached. Input validation failed.`);
         return undefined;
     }
+
+    static getStandardQpList(list:any[], prefix: string, lastModelNumber: number = 0) {
+        const items: QuickPickItem[] = [];
+        let i = lastModelNumber;
+        for (let elem of list) {
+            i++;
+            items.push({
+                label: i + ". " + prefix + elem.name,
+                description: elem.description,
+            });
+        }
+        return items;
+    } 
 }
